+++ .github/TestOutputs/CHIC_Stored.txt
