+++ .github/workflows/Benchmark.yml
@@ -21,7 +21,7 @@ jobs:
       matrix:
         include:
           - name: PlotUpdate Beam
-            cmake_options: -DUseCUDAProb3Linear=1 -DUseNuFASTLinear=1 -DUseProb3ppLinear=1 -DUseOscProb=1 -DUseNuSQUIDSLinear=1 -DUseGLoBESLinear=1 -DUseOscLibLinear=1 -DUseCHICLinear=1
+            cmake_options: -DUseCUDAProb3Linear=1 -DUseNuFASTLinear=1 -DUseProb3ppLinear=1 -DUseOscProb=1 -DUseNuSQUIDSLinear=1 -DUseGLoBESLinear=1 -DUseOscLibLinear=1 -DUseCHIC=1
             exec: DragRace
             argument: 1000 NuOscillatorConfigs/ExampleOscillationParameters.yaml NuOscillatorConfigs/Binned_CUDAProb3Linear.yaml NuOscillatorConfigs/Binned_NuFASTLinear.yaml NuOscillatorConfigs/Binned_Prob3ppLinear.yaml NuOscillatorConfigs/Binned_OscProbLinear.yaml NuOscillatorConfigs/Binned_NuSQUIDSLinear.yaml NuOscillatorConfigs/Binned_GLoBESLinear.yaml  NuOscillatorConfigs/Binned_CHICLinear.yaml NuOscillatorConfigs/Binned_OscLibLinear.yaml
             title: TimingDistributionBeam.png
