+++ .github/workflows/CIValidations.yml
@@ -30,11 +30,11 @@ jobs:
             exec: DragRace
             argument: 10 NuOscillatorConfigs/ExampleOscillationParameters.yaml NuOscillatorConfigs/Unbinned_CUDAProb3.yaml NuOscillatorConfigs/Unbinned_OscProb.yaml NuOscillatorConfigs/Unbinned_NuFASTEarth.yaml
           - name: DragRace_Linear_Binned
-            cmake_options: -DUseCUDAProb3Linear=1 -DUseNuFASTLinear=1 -DUseProb3ppLinear=1 -DUseOscProb=1 -DUseNuSQUIDSLinear=1 -DUseGLoBESLinear=1 -DUseCHICLinear=1 -DUseOscLibLinear=1
+            cmake_options: -DUseCUDAProb3Linear=1 -DUseNuFASTLinear=1 -DUseProb3ppLinear=1 -DUseOscProb=1 -DUseNuSQUIDSLinear=1 -DUseGLoBESLinear=1 -DUseCHIC=1 -DUseOscLibLinear=1
             exec: DragRace
             argument: 10 NuOscillatorConfigs/ExampleOscillationParameters.yaml NuOscillatorConfigs/Binned_CUDAProb3Linear.yaml NuOscillatorConfigs/Binned_NuFASTLinear.yaml NuOscillatorConfigs/Binned_Prob3ppLinear.yaml NuOscillatorConfigs/Binned_OscProbLinear.yaml NuOscillatorConfigs/Binned_NuSQUIDSLinear.yaml NuOscillatorConfigs/Binned_GLoBESLinear.yaml NuOscillatorConfigs/Binned_CHICLinear.yaml NuOscillatorConfigs/Binned_OscLibLinear.yaml
           - name: DragRace_Linear_Unbinned
-            cmake_options: -DUseCUDAProb3Linear=1 -DUseNuFASTLinear=1 -DUseProb3ppLinear=1 -DUseOscProb=1 -DUseNuSQUIDSLinear=1 -DUseGLoBESLinear=1 -DUseCHICLinear=1 -DUseOscLibLinear=1
+            cmake_options: -DUseCUDAProb3Linear=1 -DUseNuFASTLinear=1 -DUseProb3ppLinear=1 -DUseOscProb=1 -DUseNuSQUIDSLinear=1 -DUseGLoBESLinear=1 -DUseCHIC=1 -DUseOscLibLinear=1
             exec: DragRace
             argument: 10 NuOscillatorConfigs/ExampleOscillationParameters.yaml NuOscillatorConfigs/Unbinned_CUDAProb3Linear.yaml NuOscillatorConfigs/Unbinned_NuFASTLinear.yaml NuOscillatorConfigs/Unbinned_Prob3ppLinear.yaml NuOscillatorConfigs/Unbinned_OscProbLinear.yaml NuOscillatorConfigs/Unbinned_NuSQUIDSLinear.yaml NuOscillatorConfigs/Unbinned_GLoBESLinear.yaml NuOscillatorConfigs/Unbinned_CHICLinear.yaml NuOscillatorConfigs/Unbinned_OscLibLinear.yaml
     container:
