+++ NuOscillatorConfigs/Binned_CHIC.yaml
