+++ NuOscillatorConfigs/OscProbCalcerConfigs/CHIC.yaml
